### Fixed - Fixed a bug that users with role added with add_role were not found in search
### Upgrade Guide
In versions prior to LINE Connect 3.1.1, there was a bug where users with roles added via add_role() did not appear in search results. This issue has been fixed.

The cause was that, when searching for the target users using WP_User_Query, the search was conducted based on the "display_name" instead of the role "name," which resulted in the users not being found under the expected criteria.

This has been corrected by saving the role names and using them in searches moving forward. However, please note that the select box, which previously listed the display names, will now list the role names. As a result, the previously selected roles will not be retained.

Therefore, if any settings related to roles were configured, you will need to reselect and assign the appropriate roles.

The affected areas are as follows:

- Default target role in channel settings
- Saved Send target for each post
- Triggers where roles are specified as source conditions

### 日本語訳
LINE Connect バージョン3.1.0より以前のバージョンに、add_role()で追加したロールを持つユーザーが検索でヒットしないというバグが有り、それを修正しました。
原因は対象のユーザーをWP_User_Queryで検索する際に、ロール名の「name」ではなく表示名の「display_name」で検索してしまっていたため、条件に一致するユーザーが見つからない状態となっていました。
これを、ロール名を保存して、検索時に使用するように修正しました。
ただし、これまでは表示名を選択項目としていたセレクトボックスが、今後はロール名が選択項目となり、これまで選択していたロールの設定が維持されなくなります。
そのため、ロールに関する設定を行っていた場合、再度ロールを選択して指定し直す必要があります。
対象箇所は、以下のとおりです。
- チャネル設定のデフォルトの送信対象
- 各記事の保存済み送信対象
- ソース条件でロールを指定しているトリガー